--- Godot Project/Demos/HighLevel/HighLevel.tscn
@@ -1,24 +1,0 @@
-[gd_scene load_steps=6 format=3 uid="uid://r1q2puatsy2c"]
-
-[ext_resource type="Script" path="res://Demos/HighLevel/HighLevel.cs" id="1_7lxk5"]
-[ext_resource type="Script" path="res://addons/match_maker/MatchMaker.cs" id="2_86nxv"]
-[ext_resource type="PackedScene" uid="uid://fmfujbwoir0x" path="res://Demos/HighLevel/HighLevelPlayer.tscn" id="3_k8xtg"]
-[ext_resource type="PackedScene" uid="uid://dfrhaa6ime1jf" path="res://DebugPanel/DebugPanel.tscn" id="4_4sssx"]
-[ext_resource type="PackedScene" uid="uid://cenjhm83f2riu" path="res://ConnectionPanel/ConnectionPanel.tscn" id="5_goh70"]
-
-[node name="Main" type="Node"]
-script = ExtResource("1_7lxk5")
-playerScene = ExtResource("3_k8xtg")
-
-[node name="MatchMaker" type="Node" parent="."]
-script = ExtResource("2_86nxv")
-
-[node name="MultiplayerSpawner" type="MultiplayerSpawner" parent="."]
-_spawnable_scenes = PackedStringArray("res://Demos/HighLevel/HighLevelPlayer.tscn")
-spawn_path = NodePath("..")
-
-[node name="DebugPanel" parent="." instance=ExtResource("4_4sssx")]
-unique_name_in_owner = true
-
-[node name="ConnectionPanel" parent="." instance=ExtResource("5_goh70")]
-unique_name_in_owner = true
