+++ Godot Project/Demos/RPC/RPC.tscn
@@ -1,0 +1,73 @@
+[gd_scene load_steps=5 format=3 uid="uid://r1q2puatsy2c"]
+
+[ext_resource type="Script" path="res://Demos/RPC/RPC.cs" id="1_qxg5j"]
+[ext_resource type="Script" path="res://addons/match_maker/MatchMaker.cs" id="2_s8vu0"]
+[ext_resource type="PackedScene" uid="uid://dfrhaa6ime1jf" path="res://DebugPanel/DebugPanel.tscn" id="3_t1ovn"]
+[ext_resource type="PackedScene" uid="uid://cenjhm83f2riu" path="res://ConnectionPanel/ConnectionPanel.tscn" id="4_a5ulm"]
+
+[node name="RPC" type="Node"]
+script = ExtResource("1_qxg5j")
+
+[node name="MatchMaker" type="Node" parent="."]
+script = ExtResource("2_s8vu0")
+PrintIncomingMessagesToConsole = false
+
+[node name="Panel" type="Panel" parent="."]
+anchors_preset = 15
+anchor_right = 1.0
+anchor_bottom = 1.0
+grow_horizontal = 2
+grow_vertical = 2
+
+[node name="VBoxContainer" type="VBoxContainer" parent="Panel"]
+layout_mode = 1
+anchors_preset = 15
+anchor_right = 1.0
+anchor_bottom = 1.0
+grow_horizontal = 2
+grow_vertical = 2
+theme_override_constants/separation = 16
+alignment = 1
+
+[node name="CounterLabel" type="Label" parent="Panel/VBoxContainer"]
+layout_mode = 2
+size_flags_horizontal = 4
+text = "Counter:"
+horizontal_alignment = 1
+vertical_alignment = 1
+
+[node name="HBoxContainer" type="HBoxContainer" parent="Panel/VBoxContainer"]
+layout_mode = 2
+theme_override_constants/separation = 32
+alignment = 1
+
+[node name="MinusButton" type="Button" parent="Panel/VBoxContainer/HBoxContainer"]
+custom_minimum_size = Vector2(32, 32)
+layout_mode = 2
+size_flags_horizontal = 4
+text = "-"
+
+[node name="ActualCounterLabel" type="Label" parent="Panel/VBoxContainer/HBoxContainer"]
+unique_name_in_owner = true
+layout_mode = 2
+size_flags_horizontal = 4
+text = "0"
+horizontal_alignment = 1
+vertical_alignment = 1
+
+[node name="PlusButton" type="Button" parent="Panel/VBoxContainer/HBoxContainer"]
+custom_minimum_size = Vector2(32, 32)
+layout_mode = 2
+size_flags_horizontal = 4
+text = "+"
+
+[node name="DebugPanel" parent="." instance=ExtResource("3_t1ovn")]
+unique_name_in_owner = true
+visible = false
+
+[node name="ConnectionPanel" parent="." instance=ExtResource("4_a5ulm")]
+unique_name_in_owner = true
+visible = false
+
+[connection signal="pressed" from="Panel/VBoxContainer/HBoxContainer/MinusButton" to="." method="OnMinusButtonPressed"]
+[connection signal="pressed" from="Panel/VBoxContainer/HBoxContainer/PlusButton" to="." method="OnPlusButtonPressed"]
